+++ HISTORY.rst
@@ -5,12 +5,12 @@ History
 0.2.0 (???)
 ------------------
 
+* Add n_jobs argument using joblib parallel processing
 * Allow `cv` to take the value -1 equivalently to `LeaveOneOut()`
 * Introduce the `cv` parameter to get closer to scikit-learn API
 * Remove the `n_splits`, `shuffle` and `random_state` parameters
 * Simplify the `method` parameter
 * Fix typos in documentation and add methods descriptions in sphinx
-* Add n_jobs argument using joblib parallel processing
 
 0.1.4 (2021-05-07)
 ------------------